rename to SpatialBivariate
