+++ liana/method/sp/__init__.py
@@ -1,4 +1,4 @@
 from ...utils._spatial_neighbors import spatial_neighbors
 from ._lr_bivar import lr_bivar
-from ._bivar import bivar
+from ._SpatialBivariate import bivar
 from ._misty_constructs import genericMistyData, lrMistyData, MistyData
